requirements.txt
anyio==3.7.1
beautifulsoup4==4.12.3
bs4==0.0.2
google-api-core==2.18.0
google-auth==2.29.0
google-cloud==0.34.0
google-cloud-dataproc==5.4.0
pandas==1.5.3
prefect==2.12.1
prefect-gcp==0.5.7
pydantic>=1.10.9
Pygments==2.17.2
pyspark==3.5.1
requests==2.31.0